refactor(admin): AdminDashboard sous 300 lignes — éclatement complet (H3)
AdminDashboard.vue : 1241 -> 156 lignes. Toute la logique migrée dans des
composables ; chaque section de template devient un sous-composant présentation.
Comportement et rendu identiques (mêmes services, routes, classes CSS, textes).

Composables (src/composables/) :
- useAdminDashboard.js (93) : orchestrateur, état partagé, helpers de rôle,
  navigation, onMounted ; compose les deux suivants.
- useAdminDashboardData.js (157) : loadKlassciData / loadAnalytics /
  loadCalendarEvents (appels KLASSCI + analytics + mapping calendrier), verbatim.
- useDashboardFormatters.js (60) : getInitials / getRoleLabel / getRoleClass /
  formatDate (fonctions pures de présentation).

Sous-composants (src/components/dashboard/) : DashboardStatsCards,
DashboardActivityChart, DashboardPendingTasks, DashboardSystemWidgets,
DashboardQuickActions, DashboardRoleActions, DashboardClasses,
DashboardRecentUsers, DashboardMatieres, DashboardInfoBanner. CSS de chaque
section déplacé verbatim (y compris @media) ; les classes partagées
(.widget-card/.widget-header/.widget-icon/.widget-title, .empty-state-inline/
.empty-icon/.empty-message, .loading-indicator, .stat-label/.stat-value)
dupliquées dans chaque enfant qui les utilise (style scoped).

Tests (tests/unit/) : useAdminDashboard, useDashboardFormatters,
DashboardStatsCards, DashboardQuickActions, DashboardRoleActions — 23/23 verts.
Build vite : exit 0.

Dette préexistante conservée à l'identique (non « corrigée ») :
- Import NotTIONSWidget retiré : son unique usage (<NotificationsWidget>) était
  déjà commenté dans le template d'origine (ligne 325) — l'import était donc
  déjà mort. Le commentaire de markup est conservé tel quel. Aucune autre
  modification de comportement.
- CSS mort conservé verbatim dans les enfants concernés : .permissions-grid et
  .permission-badge n'ont AUCUN markup correspondant dans le template d'origine.
  Pour rester verbatim sans introduire de sélecteur orphelin dans un fichier
  scoped sans cible, ce bloc mort N'A PAS été reporté (il ne ciblait rien et
  aurait été supprimé par le scoping de toute façon). Signalé explicitement.
